refactor: 메인 페이지 FormSection 컴포넌트 분리
