+++ apps/frontend/src/_components/pages/main/FormSection.module.css
@@ -1,0 +1,6 @@
+.form {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
