+++ apps/frontend/src/app/main/page.module.css
@@ -1,6 +1,11 @@
 .container {
   width: 100%;
   max-width: 400px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: var(--space-4);
 
   @media (min-width: 768px) {
     padding-top: var(--layout-page-top-padding);
@@ -17,11 +22,4 @@
   @media (max-width: 767px) {
     font-size: var(--font-base) !important;
   }
-}
-
-.form {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
 }
